+++ src/Domain/ReservedSeat.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TrainReservation;
+namespace TrainReservation\Domain;
 
 final class ReservedSeat extends Seat
 {
