+++ src/Domain/TicketOffice.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TrainReservation;
+namespace TrainReservation\Domain;
 
 class TicketOffice
 {
