+++ src/Domain/TrainDataProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace TrainReservation;
+namespace TrainReservation\Domain;
 
 interface TrainDataProvider
 {
